@@ -100,7 +100,7 @@ class ApplicationController < ActionController::Base
     @batches = []
     return if action_name == "index" and controller_name == "mailbox"
     if User.current
-      @notify_number = User.current.mailbox.inbox(:unread => true).count(:id, :distinct => true)
+      @notify_number = User.current.mailbox.inbox(:unread => true).distinct(:id).count
       @batches = User.current.mailbox.inbox.map { |msg| msg.last_message.body[/<a class="batchid ui-helper-hidden">(.*)<\/a>The file(.*)/,1] }.select { |val| !val.blank? }
     end
   end
