@@ -26,7 +26,7 @@ class CollectionsController < ApplicationController
 
   def after_destroy (id)
     respond_to do |format|
-      format.html { redirect_to dashboard_path, notice: 'Collection was successfully deleted.' }
+      format.html { redirect_to sufia.dashboard_index_path, notice: 'Collection was successfully deleted.' }
       format.json { render json: {id:id}, status: :destroyed, location: @collection }
     end
   end
