@@ -1,18 +1,18 @@
 ScholarSphere::Application.routes.draw do
   resource :landing_page, only: [:new, :create]
-  match ':managedata' => 'landing_pages#new', :as => :request_info, :constraints => { :managedata => /managedata/i }
-  match ':managedata/:thankyou' => 'landing_pages#thanks', :as => :request_thanks, :constraints => { :managedata => /managedata/i, :thankyou=> /thankyou/i }
+  get ':managedata' => 'landing_pages#new', :as => :request_info, :constraints => { :managedata => /managedata/i }
+  get ':managedata/:thankyou' => 'landing_pages#thanks', :as => :request_thanks, :constraints => { :managedata => /managedata/i, :thankyou=> /thankyou/i }
 
   root :to => "catalog#index"
 
   devise_for :users
 
   # Login/logout route to destroy session
-  match 'logout' => 'sessions#destroy', :as => :destroy_user_session
-  match 'login' => 'sessions#new', :as => :new_user_session
+  get 'logout' => 'sessions#destroy', :as => :destroy_user_session
+  get 'login' => 'sessions#new', :as => :new_user_session
 
   # "Recently added files" route for catalog index view (needed before BL routes)
-  match "catalog/recent" => "catalog#recent", :as => :catalog_recent
+  get "catalog/recent" => "catalog#recent", :as => :catalog_recent
 
   Blacklight.add_routes(self)
   Hydra::BatchEdit.add_routes(self)
@@ -37,7 +37,7 @@ ScholarSphere::Application.routes.draw do
     end
     # Usage stats
     constraints StatsAdmin do
-      match 'stats' => 'stats#index', :as => :stats
+      get 'stats' => 'stats#index', :as => :stats
     end
   end
 
@@ -51,55 +51,55 @@ ScholarSphere::Application.routes.draw do
     end
   end
   
-  match '/users/:user_id/depositors' =>  'depositors#create', via: :post, as:'user_depositors'
-  match '/users/:user_id/depositors/:id' =>  'depositors#destroy', via: :delete, as:'user_depositor'
+  post '/users/:user_id/depositors' =>  'depositors#create', as:'user_depositors'
+  delete '/users/:user_id/depositors/:id' =>  'depositors#destroy', as:'user_depositor'
 
   mount Hydra::Collections::Engine => '/'
   mount Sufia::Engine => '/'
   mount HydraEditor::Engine => '/'
 
   # Downloads controller route
-  resources :downloads, :only => "show"
+  # resources :downloads, :only => "show"
 
   # Batch edit routes
-  match 'batches/:id/edit' => 'batch#edit', :as => :batch_edit
-  match 'batches/:id/' => 'batch#update', :as => :batch_generic_files
+  # get 'batches/:id/edit' => 'batch#edit', :as => :batch_edit
+  # put 'batches/:id/' => 'batch#update', :as => :batch_generic_files
 
   # adding user route here to fix routing issue not found page=nil
-  match 'users' => 'users#index', :as => :profiles
+  # get 'users' => 'users#index', :as => :profiles
 
 
 
-  # Dashboard routes (based partly on catalog routes)
-  match 'dashboard' => 'dashboard#index', :as => :dashboard
-  match 'dashboard/activity' => 'dashboard#activity', :as => :dashboard_activity
-  match 'dashboard/facet/:id' => 'dashboard#facet', :as => :dashboard_facet
+  # # Dashboard routes (based partly on catalog routes)
+  # get 'dashboard' => 'dashboard#index', :as => :dashboard
+  # get 'dashboard/activity' => 'dashboard#activity', :as => :dashboard_activity
+  # get 'dashboard/facet/:id' => 'dashboard#facet', :as => :dashboard_facet
 
   # Messages
-  match 'notifications' => 'mailbox#index', :as => :mailbox
-  match 'notifications/delete_all' => 'mailbox#delete_all', :as => :mailbox_delete_all
-  match 'notifications/:uid/delete' => 'mailbox#delete', :as => :mailbox_delete
+  # get 'notifications' => 'mailbox#index', :as => :mailbox
+  # match 'notifications/delete_all' => 'mailbox#delete_all', :as => :mailbox_delete_all
+  # match 'notifications/:uid/delete' => 'mailbox#delete', :as => :mailbox_delete
 
   # Authority vocabulary queries route
-  match 'authorities/:model/:term' => 'authorities#query'
+  # get 'authorities/:model/:term' => 'authorities#query'
 
   # Advanced search
-  match 'search' => 'advanced#index', :as => :advanced
+  # get 'search' => 'advanced#index', :as => :advanced
 
   # LDAP-related routes for group and user lookups
-  match 'directory/user/:uid' => 'directory#user'
-  match 'directory/user/:uid/:attribute' => 'directory#user_attribute'
-  match 'directory/group/:cn' => 'directory#group', :constraints => { :cn => /.*/ }
+  # get 'directory/user/:uid' => 'directory#user'
+  # get 'directory/user/:uid/:attribute' => 'directory#user_attribute'
+  # get 'directory/group/:cn' => 'directory#group', :constraints => { :cn => /.*/ }
 
   # Contact form routes
-  match 'contact' => 'contact_form#create', :via => :post, :as => :contact_form_index
-  match 'contact' => 'contact_form#new', :via => :get, :as => :contact_form_index
+  # post 'contact' => 'contact_form#create', :as => :contact_form_index
+  # get 'contact' => 'contact_form#new', :as => :contact_form_index
 
   # Static page routes (workaround)
-  match ':action' => 'static#:action', :constraints => { :action => /about|help|terms|zotero|mendeley|agreement|subject_libraries|versions/ }, :as => :static
+  # get ':action' => 'static#:action', :constraints => { :action => /about|help|terms|zotero|mendeley|agreement|subject_libraries|versions/ }, :as => :static
 
-  # Catch-all (for routing errors)
-  unless Rails.env.development? || Rails.env.test?
-    match '*error' => 'errors#routing'
-  end
+  # # Catch-all (for routing errors)
+  # unless Rails.env.development? || Rails.env.test?
+  #   match '*error' => 'errors#routing', via: [:get, :post, :put]
+  # end
 end
